pyproject.toml
[project]
name = "resume-analyser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-docx>=1.1.0",
    "flask>=3.1.2",
    "pandas>=2.3.2",
    "pypdf2>=3.0.1",
]
